Putting order_quoting and order_validation into shared crate
