+++ crates/orderbook/src/api/create_order.rs
@@ -1,7 +1,4 @@
-use crate::{
-    order_validation::{PartialValidationError, ValidationError},
-    orderbook::{AddOrderError, Orderbook},
-};
+use crate::orderbook::{AddOrderError, Orderbook};
 use anyhow::Result;
 use model::order::{OrderCreation, OrderUid};
 use shared::api::{error, extract_payload, internal_error, ApiReply, IntoWarpReply};
@@ -14,164 +11,6 @@ pub fn create_order_request() -> impl Filter<Extract = (OrderCreation,), Error =
     warp::path!("orders")
         .and(warp::post())
         .and(extract_payload())
-}
-
-impl IntoWarpReply for PartialValidationError {
-    fn into_warp_reply(self) -> ApiReply {
-        match self {
-            Self::UnsupportedBuyTokenDestination(dest) => with_status(
-                error("UnsupportedBuyTokenDestination", format!("Type {dest:?}")),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::UnsupportedSellTokenSource(src) => with_status(
-                error("UnsupportedSellTokenSource", format!("Type {src:?}")),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::UnsupportedOrderType => with_status(
-                error(
-                    "UnsupportedOrderType",
-                    "This order type is currently not supported",
-                ),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::Forbidden => with_status(
-                error("Forbidden", "Forbidden, your account is deny-listed"),
-                StatusCode::FORBIDDEN,
-            ),
-            Self::InsufficientValidTo => with_status(
-                error(
-                    "InsufficientValidTo",
-                    "validTo is not far enough in the future",
-                ),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::ExcessiveValidTo => with_status(
-                error("ExcessiveValidTo", "validTo is too far into the future"),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::TransferEthToContract => with_status(
-                error(
-                    "TransferEthToContract",
-                    "Sending Ether to smart contract wallets is currently not supported",
-                ),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::InvalidNativeSellToken => with_status(
-                error(
-                    "InvalidNativeSellToken",
-                    "The chain's native token (Ether/xDai) cannot be used as the sell token",
-                ),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::SameBuyAndSellToken => with_status(
-                error(
-                    "SameBuyAndSellToken",
-                    "Buy token is the same as the sell token.",
-                ),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::UnsupportedSignature => with_status(
-                error("UnsupportedSignature", "signing scheme is not supported"),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::UnsupportedToken(token) => with_status(
-                error("UnsupportedToken", format!("Token address {token:?}")),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::Other(err) => with_status(
-                internal_error(err.context("partial_validation")),
-                StatusCode::INTERNAL_SERVER_ERROR,
-            ),
-        }
-    }
-}
-
-impl IntoWarpReply for ValidationError {
-    fn into_warp_reply(self) -> ApiReply {
-        match self {
-            Self::Partial(pre) => pre.into_warp_reply(),
-            Self::QuoteNotFound => with_status(
-                error(
-                    "QuoteNotFound",
-                    "could not find quote with the specified ID",
-                ),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::InvalidQuote => with_status(
-                error(
-                    "InvalidQuote",
-                    "the quote with the specified ID does not match the order",
-                ),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::PriceForQuote(err) => err.into_warp_reply(),
-            Self::MissingFrom => with_status(
-                error(
-                    "MissingFrom",
-                    "From address must be specified for on-chain signature",
-                ),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::WrongOwner(owner) => with_status(
-                error(
-                    "WrongOwner",
-                    format!("Address recovered from signature {owner} does not match from address"),
-                ),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::InsufficientBalance => with_status(
-                error(
-                    "InsufficientBalance",
-                    "order owner must have funds worth at least x in his account",
-                ),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::InsufficientAllowance => with_status(
-                error(
-                    "InsufficientAllowance",
-                    "order owner must give allowance to VaultRelayer",
-                ),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::InvalidSignature => with_status(
-                error("InvalidSignature", "invalid signature"),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::InsufficientFee => with_status(
-                error("InsufficientFee", "Order does not include sufficient fee"),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::SellAmountOverflow => with_status(
-                error(
-                    "SellAmountOverflow",
-                    "Sell amount + fee amount must fit in U256",
-                ),
-                StatusCode::INTERNAL_SERVER_ERROR,
-            ),
-            Self::TransferSimulationFailed => with_status(
-                error(
-                    "TransferSimulationFailed",
-                    "sell token cannot be transferred",
-                ),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::ZeroAmount => with_status(
-                error("ZeroAmount", "Buy or sell amount is zero."),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::IncompatibleSigningScheme => with_status(
-                error(
-                    "IncompatibleSigningScheme",
-                    "Signing scheme is not compatible with order placement method.",
-                ),
-                StatusCode::BAD_REQUEST,
-            ),
-            Self::Other(err) => with_status(
-                internal_error(err.context("order_validation")),
-                StatusCode::INTERNAL_SERVER_ERROR,
-            ),
-        }
-    }
 }
 
 impl IntoWarpReply for AddOrderError {
