+++ crates/orderbook/src/database/orders.rs
@@ -1,5 +1,4 @@
 use super::Postgres;
-use crate::order_quoting::Quote;
 use anyhow::{anyhow, Context as _, Result};
 use chrono::{DateTime, Utc};
 use database::{
@@ -22,6 +21,7 @@ use model::{
 use num::Zero;
 use number_conversions::{big_decimal_to_big_uint, big_decimal_to_u256, u256_to_big_decimal};
 use primitive_types::H160;
+use shared::order_quoting::Quote;
 use sqlx::{types::BigDecimal, Connection, PgConnection};
 use std::convert::TryInto;
