+++ crates/shared/Cargo.toml
@@ -18,6 +18,7 @@ cached = { version = "0.34", default-features = false }
 chrono = { version = "0.4", default-features = false, features = ["clock"] }
 clap = { version = "3.1", features = ["derive", "env"] }
 contracts = { path = "../contracts" }
+database = { path = "../database" }
 derivative = "2.2"
 ethcontract = { version = "0.19.0", default-features = false }
 ethcontract-mock = { version = "0.19.0", default-features = false }
@@ -43,6 +44,7 @@ scopeguard = "1.1.0"
 serde = "1.0"
 serde_json = "1.0"
 serde_with = { version = "1.11", default-features = false }
+secp256k1 = "0.21"
 thiserror = "1.0"
 time = { version = "0.3", features = ["macros"] }
 tokio = { version = "1.15", features = ["macros", "time"] }
