+++ crates/shared/src/fee_subsidy/cow_token.rs
@@ -1,10 +1,10 @@
 use super::{FeeSubsidizing, Subsidy, SubsidyParameters};
+use crate::transport::buffered::{Buffered, Configuration};
 use anyhow::{Context, Result};
 use cached::{Cached, TimedSizedCache};
 use contracts::{CowProtocolToken, CowProtocolVirtualToken};
 use ethcontract::Web3;
 use primitive_types::{H160, U256};
-use shared::transport::buffered::{Buffered, Configuration};
 use std::collections::BTreeMap;
 use std::{sync::Mutex, time::Duration};
 
@@ -131,14 +131,14 @@ impl FeeSubsidizing for CowSubsidy {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::Web3;
     use hex_literal::hex;
-    use shared::Web3;
 
     #[tokio::test]
     #[ignore]
     async fn mainnet() {
-        shared::tracing::initialize_for_tests("orderbook=debug");
-        let transport = shared::transport::create_env_test_transport();
+        crate::tracing::initialize_for_tests("orderbook=debug");
+        let transport = crate::transport::create_env_test_transport();
         let web3 = Web3::new(transport);
         let token = CowProtocolToken::deployed(&web3).await.unwrap();
         let vtoken = CowProtocolVirtualToken::deployed(&web3).await.unwrap();
