+++ crates/shared/src/lib.rs
@@ -11,6 +11,7 @@ pub mod conversions;
 pub mod current_block;
 pub mod ethcontract_error;
 pub mod event_handling;
+pub mod fee_subsidy;
 pub mod gas_price;
 pub mod gas_price_estimation;
 pub mod http_client;
@@ -19,6 +20,8 @@ pub mod maintenance;
 pub mod metrics;
 pub mod network;
 pub mod oneinch_api;
+pub mod order_quoting;
+pub mod order_validation;
 pub mod paraswap_api;
 pub mod price_estimation;
 pub mod rate_limiter;
